feat(libdelfos): Phase 4 — nanobind Python bindings + NativeGraphStore
Implements the Python extension module and NativeGraphStore (Phase 4 of
docs/libdelfos-plan.md).

New files:
  libdelfos/bindings/CMakeLists.txt   — nanobind_add_module, install to delfos/
  libdelfos/bindings/py_delfos.cpp    — Store C++ class (Graph+VectorIndex+
                                        manifest+txn) + NB_MODULE(_delfos)
  delfos/store/native_store.py        — NativeGraphStore (GraphStore ABC)
  delfos/_delfos.pyi                  — type stubs for pyright strict mode
  tests/store/test_native_store.py    — 28 parity tests (mirrors test_duckdb_store)

Modified:
  CMakeLists.txt         — SKBUILD detection (forces DELFOS_BUILD_BINDINGS ON
                           for pip builds); nanobind via Python pkg or FetchContent
  pyproject.toml         — switch build backend to scikit-build-core; add
                           CMAKE_CXX_COMPILER=clang++ for wheel builds
  libdelfos/include/delfos/node.hpp       — embedding: float → double (to match
                                            DuckDB float64 precision for round-trips)
  libdelfos/include/delfos/vector_index.hpp — add double overloads for insert()/search()
  libdelfos/include/delfos/graph.hpp      — add nodes_for_file() accessor
  libdelfos/flatbuffers/delfos.fbs        — embedding: [float] → [double]
  libdelfos/flatbuffers/delfos_generated.h — regenerated with double embedding
  libdelfos/tests/test_snapshot.cpp       — update vec() helper to return double

Phase 4 acceptance criteria met:
  - from delfos._delfos import Store, NodeData succeeds (pip install -e .)
  - NativeGraphStore implements all GraphStore ABC methods
  - 28/28 parity tests pass (same coverage as test_duckdb_store.py)
  - pyright --strict: 0 errors on native_store.py (1 expected warning for
    native module source resolution, not a type error)
  - uv build produces delfos-0.1.0-cp312-cp312-linux_x86_64.whl
  - 76/76 C++ tests + 56/56 Python tests pass
  - ruff: all checks pass

+++ libdelfos/tests/test_snapshot.cpp
@@ -12,10 +12,10 @@ using namespace delfos;
 static constexpr std::size_t DIM = 4;
 static constexpr char MODEL[]    = "test-model";
 
-static std::vector<float> vec(float v0, float v1, float v2, float v3) {
+static std::vector<double> vec(double v0, double v1, double v2, double v3) {
     // normalise so cosine similarity is well-defined
-    float len = std::sqrt(v0*v0 + v1*v1 + v2*v2 + v3*v3);
-    if (len == 0.0f) len = 1.0f;
+    double len = std::sqrt(v0*v0 + v1*v1 + v2*v2 + v3*v3);
+    if (len == 0.0) len = 1.0;
     return {v0/len, v1/len, v2/len, v3/len};
 }
